fix: fuzzy selection prompt
The prompt now correctly displays the rights prefix' and suffix'
depending on the current state of the prompt. The goal was to create a
prompt that does render its text followed by the user input. If no
prompt is present, the user input is still rendered.

To successfully implement this behavior, I had to extend the
`Theme` trait for an additional method called `format_fuzzy_prompt`, as
there was no way to implement the above behavior with the currently
available methods. The method does take an `Option<&str>` as its prompt
type instead of a `&str` as all other methods. This is due to the fact
that the user input should get rendered even if no prompt has been
specified.

An alternative approach would have been to render the two separately,
but I couldn't find a way to append to the current prompt.