Simulate cursor in search term input prompt
We style the current position in the search term to simulate the cursor, so the user can see where the cursor is at.
Currently, the cursor is fixed to black font on white background, but it should be styleable in the future.

If the cursor is at the end of the search term, we print the whole search term followed by a string containing a space to simulate the cursor at the end of the prompt