Feature: Add fuzzy single select prompt (#124)
* feat: add fuzzy single select prompt

* fix: fuzzy selection prompt

The prompt now correctly displays the rights prefix' and suffix'
depending on the current state of the prompt. The goal was to create a
prompt that does render its text followed by the user input. If no
prompt is present, the user input is still rendered.

To successfully implement this behavior, I had to extend the
`Theme` trait for an additional method called `format_fuzzy_prompt`, as
there was no way to implement the above behavior with the currently
available methods. The method does take an `Option<&str>` as its prompt
type instead of a `&str` as all other methods. This is due to the fact
that the user input should get rendered even if no prompt has been
specified.

An alternative approach would have been to render the two separately,
but I couldn't find a way to append to the current prompt.

* chore: apply requested changes

* Add fuzzy-matcher to Cargo.toml

* Fix missing feature mapping for fuzzy format function

* Add missing feature mapping to fuzzy select prompt formatting function

* Remove unnecessary `Option`

* Add doc to FuzzySelect

* Make `item` and `with_prompt` functions generic

* Apply suggestions for interact functions from PR #79

* Apply suggested pages calculation from #79

* Use map instead of match

* Use term.flush() instead of clear_last_lines(filtered_list.len())?;

* Merge imports

* Add commas to match in fuzzy loop

* Fix all build errors

* Correctly determine position of selection in items array to return usize

* fmt in fuzzy_select

* Fix fuzzyselect example to work with usize

* Re-add clearing of last lines to clear

* Use Arrow keys for navigating the input text

* Make stuff prettier and add what's left to do as comments

* Flush after selecting prompt item

* Rename case sensitive property for more clarity what it is about

* Fix arrow down matching crash when filtered list is empty

* Remove Option wrap from prompt

* Remove unnecessary writing of tail which caused weird printing

Tail method was taken from Input, where a Vec<char> is used
We work with a string here, so no need to write the tail

* Fix typo

* Add missing term flushes

* Add hide and show (console) cursor commands

* Add clear to the end of the rendering loop

* Mark pages variable unused until paging is solved

* Simulate cursor in search term input prompt

We style the current position in the search term to simulate the cursor, so the user can see where the cursor is at.
Currently, the cursor is fixed to black font on white background, but it should be styleable in the future.

If the cursor is at the end of the search term, we print the whole search term followed by a string containing a space to simulate the cursor at the end of the prompt

* Use correct cursor calculation and display in Theme trait

* Make cursor styleable

* Add changes suggested by @pksunkara

* Remove case ignoring, since the `fuzzy-matcher` library already handles that

* Rename "position" variable in example to "selection"

* Remove paging from fuzzy-select

Co-authored-by: Sven Lechner <SirWindfield@users.noreply.github.com>