Fix crash in fuzzy select
Currently when using the fuzzy select without any defaults, the default
would be set to `!0`. This is problematic since this default value is
used upon pressing enter to index into the provided options and cause an
out of bounds.

This patch fixes this issue by using a proper `Option` rather than
encoding a missing default into an integer, allowing for distinct
treatment of the scenario where the selection has not been set yet.