Handle executable with spaces in name
If the user provides an executable like `'/path/to/editor with spaces' arg1 arg2`, we will mistakenly detect the editor executable as `'/path/to/editor`, and the following words will be interpreted as arguments. This commit lexes the string as done by the POSIX shell to identify the executable in the case that it is multiple words.