Added recordings of examples to the readme
When programmers search for a great library they often find multiple and need to decide which one is worth being used.
A good readme showing the possibilities of a library helps them to quickly identify if it matches their situation.
Forcing them to download the library and to run the examples is unnecessarily anoying