feat: add disabled items support in multiselect
Sometimes it's nice to show a user a list of options, some of which the
user cannot interact with.

e.g. A list of linux users on this machine to import to a new machine, I
would like to show root but without allowing the user to deselect it.

Current implementation limitations:

- No theme support - there's no indication that an item is disabled
  until the user attempts to interact with it. Would adding such visual
  indication be a good idea from a compatibility perspective? Are themes
  all built-in or possibly user defined?

- Cannot dynamically add items with a different disabled state. This
  would require modifying the signature of e.g. `item_checked` which
  would be a breaking change. I could also add a new method but I wasn't
  sure if it's worth it.

- No tests

- No documentation

Do you think this is a desired feature? If so, I can work on improving
the limitations. Should this maybe be a separate widget altogether?